Update README to enhance project clarity and usability
Providing an improved overview of the project by merging information from the AnimeFLV and JKAnime modules clarifies functionality. The new structure outlines key components and methods, making it easier for developers to understand and utilize the project effectively.

Emphasized installation dependencies and included usage examples to assist new users in getting started quickly. Additionally, a reminder to comply with the terms of service for scraping is now highlighted, ensuring ethical usage of the project.